Move basic mode into a new Django app
