+++ src/lgr_web/urls.py
@@ -7,6 +7,8 @@ The `urlpatterns` list routes URLs to views. For more information please see:
 """
 from __future__ import unicode_literals
 from django.conf.urls import include, url
+
+import lgr_basic.urls
 import lgr_editor.urls
 import lgr_validator.urls
 import lgr_tools.urls
@@ -18,6 +20,7 @@ urlpatterns = [
     url(r'^validator/', include(lgr_validator.urls.urlpatterns)),
     url(r'^tools/', include(lgr_tools.urls.urlpatterns)),
     url(r'^render/', include(lgr_renderer.urls.urlpatterns)),
+    url(r'^basic/', include(lgr_basic.urls.urlpatterns)),
     url(r'^i18n/', include('django.conf.urls.i18n')),
     url(r'^$', views.index, name='homepage'),
 ]
